admin access check

+++ controllers/eventcontroller.js
@@ -8,6 +8,13 @@ const Prisma = new PrismaClient();
 //this controller is used to add event, only admin can add event
 export const addEvent = async (req, res) => {
   try {
+    const isAdmin = await Prisma.admin.findUnique({
+      where: { id: req.body.userId },
+    });
+    
+    if (!isAdmin) {
+      return res.status(403).json({ message: "Access denied. Only admins can access this route." });
+    }
     const validation = addEventSchema.safeParse(req.body);
     if (!validation.success) {
     return res.status(400)
@@ -35,6 +42,13 @@ export const addEvent = async (req, res) => {
 //this controller is used to delete event by id, only admin can delete event
 export const deleteEventById = async (req, res) => {
     try {
+      const isAdmin = await Prisma.admin.findUnique({
+        where: { id: req.body.userId },
+      });
+      
+      if (!isAdmin) {
+        return res.status(403).json({ message: "Access denied. Only admins can access this route." });
+      }
       const { eventId } = req.params;
       await Prisma.event.delete({
         where: {
@@ -51,6 +65,13 @@ export const deleteEventById = async (req, res) => {
   //this controller is used to update event by id, only admin can update event
 export const updateEventById = async (req, res) => {
     try {
+      const isAdmin = await Prisma.admin.findUnique({
+        where: { id: req.body.userId },
+      });
+      
+      if (!isAdmin) {
+        return res.status(403).json({ message: "Access denied. Only admins can access this route." });
+      }
       const { eventId } = req.params;
       const validation = addEventSchema.safeParse(req.body);
       if (!validation.success) {
